+++ @xen-orchestra/web/src/modules/traffic-rules/components/actions/TrafficRuleActions.vue
@@ -6,7 +6,7 @@
     left-icon="action:edit"
     :disabled="!canEditTrafficRule"
     :busy="isEditingTrafficRule"
-    @click="openTrafficRuleEditDrawer()"
+    @click="editTrafficRule()"
   >
     {{ t('action:edit') }}
   </UiButton>
@@ -15,8 +15,7 @@
 
 <script lang="ts" setup>
 import { useTrafficRuleDelete } from '@/modules/traffic-rules/composables/use-traffic-rule-delete.composable.ts'
-import { useTrafficRuleDeleteModal } from '@/modules/traffic-rules/composables/use-traffic-rule-delete-modal.composable.ts'
-import { useTrafficRuleEditDrawer } from '@/modules/traffic-rules/composables/use-traffic-rule-edit-drawer.composable.ts'
+import { useTrafficRuleEdit } from '@/modules/traffic-rules/composables/use-traffic-rule-edit.composable.ts'
 import VtsDeleteButton from '@core/components/delete-button/VtsDeleteButton.vue'
 import UiButton from '@core/components/ui/button/UiButton.vue'
 import type { TrafficRule } from '@vates/types'
@@ -26,19 +25,9 @@ const { rule } = defineProps<{
   rule: TrafficRule
 }>()
 
+const { t } = useI18n()
 
 const { deleteTrafficRules, canDeleteTrafficRules, isDeletingTrafficRules } = useTrafficRuleDelete(() => [rule])
 
-
-const {
-  openModal: openTrafficRuleDeleteModal,
-  canRun: canDeleteTrafficRule,
-  isRunning: isDeletingTrafficRule,
-} = useTrafficRuleDeleteModal(() => [rule])
-
-const {
-  openDrawer: openTrafficRuleEditDrawer,
-  isRunning: isEditingTrafficRule,
-  canRun: canEditTrafficRule,
-} = useTrafficRuleEditDrawer(() => rule)
+const { editTrafficRule, canEditTrafficRule, isEditingTrafficRule } = useTrafficRuleEdit(() => rule)
 </script>
